requirements.txt
azure-core==1.23.0
azure-data-tables==12.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
microsoftgraph-python==1.1.1
msrest==0.6.21
oauthlib==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
zipp==3.7.0
